requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
appdirs==1.4.4
    # via black
astroid==2.5.1
    # via pylint
black==20.8b1
    # via -r requirements.in
cassandra-driver==3.24.0
    # via -r requirements.in
click==7.1.2
    # via
    #   -r requirements.in
    #   black
    #   geomet
    #   rq
geomet==0.2.1.post1
    # via cassandra-driver
isort==5.7.0
    # via
    #   -r requirements.in
    #   pylint
lazy-object-proxy==1.5.2
    # via astroid
mccabe==0.6.1
    # via pylint
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
mypy==0.812
    # via -r requirements.in
pathspec==0.8.1
    # via black
pylint==2.7.2
    # via -r requirements.in
redis==3.5.3
    # via
    #   -r requirements.in
    #   rq
regex==2020.11.13
    # via black
rq==1.7.0
    # via -r requirements.in
six==1.15.0
    # via
    #   cassandra-driver
    #   geomet
toml==0.10.2
    # via
    #   black
    #   pylint
typed-ast==1.4.2
    # via
    #   black
    #   mypy
typing-extensions==3.7.4.3
    # via
    #   black
    #   mypy
wrapt==1.12.1
    # via astroid
